Force-moved the files (renaming did not work for some reasons)
